Add main part home page
-- Prepare Comments button
-- Prepare reservation button
-- Create html template
-- Add API call
