--- src/add_vue.js
@@ -1,14 +1,0 @@
-// Shorthand function for creating a DOM element
-// elem = string, classes = array of string(s), parent = DOM element
-const addElem = (elem, classes, parent) => {
-    const createdElem = document.createElement(elem);
-    if (classes !== undefined) {
-      classes.forEach((cl) => createdElem.classList.add(cl));
-    }
-    parent.appendChild(createdElem);
-  
-    return createdElem;
-  };
-  
-  export default addElem;
-
