+++ src/template.html
@@ -39,7 +39,7 @@
   
   <section class="product">
     <div class="pro-container" id="product">
-    
+        <!-- Main DOM here -->
     </div>
   </section>
